Fixed bug with caching HLT things
+ re-organised configuration for data
